extract on IntUiBridgeText
