+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridge.kt
@@ -1,20 +1,14 @@
 package org.jetbrains.jewel.bridge.theme
 
 import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.ui.text.ExperimentalTextApi
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.platform.asComposeFontFamily
-import androidx.compose.ui.unit.sp
 import com.intellij.ide.ui.laf.darcula.DarculaUIUtil
-import com.intellij.openapi.editor.colors.ColorKey
-import com.intellij.openapi.editor.colors.EditorFontType
 import com.intellij.ui.JBColor
 import com.intellij.util.ui.DirProvider
+import javax.swing.UIManager
 import org.jetbrains.jewel.bridge.dp
 import org.jetbrains.jewel.bridge.lafName
 import org.jetbrains.jewel.bridge.readFromLaF
-import org.jetbrains.jewel.bridge.retrieveEditorColorScheme
-import org.jetbrains.jewel.bridge.retrieveTextStyle
 import org.jetbrains.jewel.bridge.toComposeColor
 import org.jetbrains.jewel.foundation.GlobalColors
 import org.jetbrains.jewel.foundation.GlobalMetrics
@@ -24,7 +18,6 @@ import org.jetbrains.jewel.foundation.theme.ThemeIconData
 import org.jetbrains.jewel.foundation.util.JewelLogger
 import org.jetbrains.jewel.ui.ComponentStyling
 import org.jetbrains.jewel.ui.DefaultComponentStyling
-import javax.swing.UIManager
 
 private val logger = JewelLogger.getInstance("JewelIntUiBridge")
 
@@ -39,43 +32,6 @@ internal fun createBridgeThemeDefinition(): ThemeDefinition {
     val editorTextStyle = retrieveEditorTextStyle()
     val consoleTextStyle = retrieveConsoleTextStyle()
     return createBridgeThemeDefinition(textStyle, editorTextStyle, consoleTextStyle)
-}
-
-public fun retrieveDefaultTextStyle(): TextStyle = retrieveTextStyle("Label.font", "Label.foreground")
-
-@OptIn(ExperimentalTextApi::class)
-public fun retrieveEditorTextStyle(): TextStyle {
-    val editorColorScheme = retrieveEditorColorScheme()
-
-    val fontSize = editorColorScheme.editorFontSize.sp
-    return retrieveDefaultTextStyle()
-        .copy(
-            color = editorColorScheme.defaultForeground.toComposeColor(),
-            fontFamily = editorColorScheme.getFont(EditorFontType.PLAIN).asComposeFontFamily(),
-            fontSize = fontSize,
-            lineHeight = fontSize * editorColorScheme.lineSpacing,
-            fontFeatureSettings = if (!editorColorScheme.isUseLigatures) "liga 0" else "liga 1",
-        )
-}
-
-@OptIn(ExperimentalTextApi::class)
-public fun retrieveConsoleTextStyle(): TextStyle {
-    val editorColorScheme = retrieveEditorColorScheme()
-    if (editorColorScheme.isUseEditorFontPreferencesInConsole) return retrieveEditorTextStyle()
-
-    val fontSize = editorColorScheme.consoleFontSize.sp
-    val fontColor =
-        editorColorScheme.getColor(ColorKey.createColorKey("BLOCK_TERMINAL_DEFAULT_FOREGROUND"))
-            ?: editorColorScheme.defaultForeground
-
-    return retrieveDefaultTextStyle()
-        .copy(
-            color = fontColor.toComposeColor(),
-            fontFamily = editorColorScheme.getFont(EditorFontType.CONSOLE_PLAIN).asComposeFontFamily(),
-            fontSize = fontSize,
-            lineHeight = fontSize * editorColorScheme.lineSpacing,
-            fontFeatureSettings = if (!editorColorScheme.isUseLigatures) "liga 0" else "liga 1",
-        )
 }
 
 internal val isDark: Boolean
